cmd/awl-tray: move dialogs-related code to separate file, simplify function signatures
